requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
networkx>=2.6.0
scikit-learn>=0.24.2
nibabel>=3.2.0  # for fMRI data handling
flask>=2.0.0
flask-cors>=3.0.10
python-dotenv>=0.19.0
transformers>=4.30.0  # for text analysis
librosa>=0.10.0  # for audio processing
opencv-python>=4.8.0  # for video processing
mediapipe>=0.10.0  # for facial landmark detection
openai>=0.27.0  # for GPT integration
pydantic>=2.0.0  # for request/response validation
python-multipart>=0.0.6  # for handling file uploads
gunicorn>=20.1.0  # for production deployment
